Logging updates to html and css docs
Have started working heavily on CSS doc and made
various edits and adjustments as needed to html.
currently struggling to get some of the box sizing
to do what I want it to do properly. will need to
start revisiting lessons and google a lot to
continue progressing through project.